pyproject.toml
[project]
name = "kphtools"
version = "0.1.0"
description = "Toolkits for KPH dynamic data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic",
    "httpx",
    "mcp",
    "openai",
    "pefile",
    "pyyaml",
    "requests",
    "signify",
]

[tool.uv]
package = false
